remove junk code
