+++ src/API/coinApi.js
@@ -1,8 +1,8 @@
 import axios from "axios";
-import { coinsApiCall } from "../Redux/Actions/coinAction";
+import { coinsApiCall, ErrorAction } from "../Redux/Actions/coinAction";
+import { apiKey } from "./key/key";
 
 const baseUrl = "https://api.coinranking.com/v2/coins";
-const apiKey = "coinranking369971eebd30ea4d94a91d301bd9fb9099e6792808fd718c";
 
 export const getData = () => async (dispatch) => {
   try {
@@ -14,6 +14,7 @@ export const getData = () => async (dispatch) => {
 
     dispatch(coinsApiCall(data));
   } catch (error) {
-    console.log(error);
+    const errorMsg = error;
+    dispatch(ErrorAction(errorMsg));
   }
 };
