+++ src/API/key/key.js
@@ -1,0 +1,2 @@
+export const apiKey =
+  "coinranking369971eebd30ea4d94a91d301bd9fb9099e6792808fd718c";
