+++ src/Redux/Reducers/coinReducer.js
@@ -1,15 +1,8 @@
-import {
-  COINS_API,
-  FAVORITE_DATA,
-  FORM_INPUT,
-  LOADER,
-  ERROR,
-} from "../Types/types";
+import { COINS_API, FAVORITE_DATA, ERROR } from "../Types/types";
 
 const initialState = {
   coinsData: [],
   favoriteData: [],
-  formInput: "",
   loading: false,
   error: "",
 };
@@ -24,10 +17,6 @@ export const coinReducer = (state = initialState, action) => {
         favoriteData: [...state.favoriteData, action.payload],
         loading: false,
       };
-    case FORM_INPUT:
-      return { ...state, formInput: action.payload };
-    case LOADER:
-      return { ...state, loading: true };
     case ERROR:
       return { ...state, error: action.payload, loading: false, coinsData: [] };
     default:
