+++ src/Redux/Types/types.js
@@ -1,5 +1,3 @@
 export const COINS_API = "COINSAPI";
 export const FAVORITE_DATA = "FAVORITE_DATA";
-export const FORM_INPUT = "FORM_INPUT";
-export const LOADER = "LOADER";
 export const ERROR = "ERROR";
